requirements.txt
fastapi
uvicorn
redis
PyYAML
